Use np.ones() with the same dtype as the field array
NumPy 2.0 raises an OverflowError instead of casting to dtype np.object_
as in previous versions. This can happen here if field.charactersitic is
to large to fit in a np.int64.

+++ src/galois/_fields/_array.py
@@ -988,7 +988,7 @@ class FieldArray(Array, metaclass=FieldArrayMeta):
         if x.ndim == 0:
             order = 1 if x == 0 else field.characteristic
         else:
-            order = field.characteristic * np.ones(x.shape, dtype=np.int64)
+            order = field.characteristic * np.ones(x.shape, dtype=x.dtype)
             order[np.where(x == 0)] = 1
 
         return order
